+++ AprgAlgebra/AprgAlgebra/src/Algebra/Constructs/ConstructUtilities.cpp
@@ -277,16 +277,15 @@ void createTermRaiseToTermsFromMultiplicationAndDivisionExpression(
     }
 
     if (shouldContinueToGetCommonExponent) {
-        for (unsigned int i = 0; i < constantFactorsOfExponents.size() && i < nonConstantFactorsOfExponents.size() &&
-                                 i < originalBases.size();
+        for (int i = 0;
+             i < static_cast<int>(constantFactorsOfExponents.size()) &&
+             i < static_cast<int>(nonConstantFactorsOfExponents.size()) && i < static_cast<int>(originalBases.size());
              i++) {
             AlbaNumber uniqueConstantExponent(constantFactorsOfExponents.at(i) / constantGcf);
-            TermsRaiseToNumbers remainingNonConstantFactors(nonConstantFactorsOfExponents.at(i));
-            remainingNonConstantFactors.subtractExponents(commonNonConstantFactorsOfExponents);
+            TermsRaiseToNumbers remainingNonConstantFactors(nonConstantFactorsOfExponents.at(i));            remainingNonConstantFactors.subtractExponents(commonNonConstantFactorsOfExponents);
             remainingNonConstantFactors.simplify();
             Term uniqueExponent(Term(uniqueConstantExponent) * remainingNonConstantFactors.getCombinedTerm());
-            uniqueExponent.simplify();
-            TermRaiseToTerms originalBaseToUniqueExponent(originalBases.at(i), uniqueExponent);
+            uniqueExponent.simplify();            TermRaiseToTerms originalBaseToUniqueExponent(originalBases.at(i), uniqueExponent);
             finalBasesWithDetails.emplace_back(
                 originalBaseToUniqueExponent.getCombinedTerm(), TermAssociationType::Positive);
         }
