+++ AprgAlgebra/AprgAlgebra/src/Algebra/Constructs/PolynomialOverPolynomial.cpp
@@ -87,18 +87,16 @@ bool PolynomialOverPolynomial::shouldPerformFactorization() const {
 }
 
 void PolynomialOverPolynomial::convertFractionCoefficientsToInteger() {
-    unsigned int numeratorMultiplier(getLcmForDenominatorCoefficients(m_numerator));
+    int numeratorMultiplier(getLcmForDenominatorCoefficients(m_numerator));
     m_numerator.multiplyNumber(numeratorMultiplier);
     m_denominator.multiplyNumber(numeratorMultiplier);
-    unsigned int denominatorMultiplier(getLcmForDenominatorCoefficients(m_denominator));
+    int denominatorMultiplier(getLcmForDenominatorCoefficients(m_denominator));
     m_numerator.multiplyNumber(denominatorMultiplier);
     m_denominator.multiplyNumber(denominatorMultiplier);
 }
-
 void PolynomialOverPolynomial::convertNegativeExponentsToPositive() {
     Monomial monomialExponentNumerator(getMonomialWithMaxNegativeExponentsAndConvertItToPositive(m_numerator));
-    Monomial monomialExponentDenominator(getMonomialWithMaxNegativeExponentsAndConvertItToPositive(m_denominator));
-    m_numerator.multiplyMonomial(monomialExponentNumerator);
+    Monomial monomialExponentDenominator(getMonomialWithMaxNegativeExponentsAndConvertItToPositive(m_denominator));    m_numerator.multiplyMonomial(monomialExponentNumerator);
     m_numerator.multiplyMonomial(monomialExponentDenominator);
     m_denominator.multiplyMonomial(monomialExponentNumerator);
     m_denominator.multiplyMonomial(monomialExponentDenominator);
@@ -142,21 +140,19 @@ void PolynomialOverPolynomial::factorizeRemoveCommonFactorsInNumeratorAndDenomin
     }
 }
 
-unsigned int PolynomialOverPolynomial::getLcmForDenominatorCoefficients(Polynomial const& polynomial) {
-    unsigned int lcm(1);
+int PolynomialOverPolynomial::getLcmForDenominatorCoefficients(Polynomial const& polynomial) {
+    int lcm(1);
     for (Monomial const& monomial : polynomial.getMonomialsConstReference()) {
         AlbaNumber const& coefficient(monomial.getConstantConstReference());
         if (coefficient.isFractionType()) {
             AlbaNumber::FractionData fractionData(coefficient.getFractionData());
-            lcm = getLeastCommonMultiple(lcm, fractionData.denominator);
+            lcm = getLeastCommonMultiple(lcm, static_cast<int>(fractionData.denominator));
         }
     }
-    return lcm;
-}
+    return lcm;}
 
 Monomial PolynomialOverPolynomial::getMonomialWithMaxNegativeExponentsAndConvertItToPositive(
-    Polynomial const& polynomial) {
-    Monomial resultMonomial(1, {});
+    Polynomial const& polynomial) {    Monomial resultMonomial(1, {});
     Monomial::VariablesToExponentsMap const& resultVariableMap(
         resultMonomial.getVariablesToExponentsMapConstReference());
     for (Monomial const& monomial : polynomial.getMonomialsConstReference()) {
