+++ AprgAlgebra/AprgAlgebra/src/Algebra/Solution/Solver/SolverUsingSubstitution/SolverUsingSubstitution.cpp
@@ -95,20 +95,18 @@ void SolverUsingSubstitution::calculateSolutions(Equations const& equations) {
 
 void SolverUsingSubstitution::calculateASolutionForAllVariables(
     MultipleVariableSolutionSet& solutionSet, Equations const& equations) {
-    unsigned int previousNumberOfVariables(0);
+    int previousNumberOfVariables(0);
     VariableNamesSet variablesWithSolution;
     do {
         previousNumberOfVariables = variablesWithSolution.size();
         calculateASolutionForOneVariable(solutionSet, equations);
         variablesWithSolution = solutionSet.getVariableNames();
-    } while (previousNumberOfVariables != variablesWithSolution.size() &&
+    } while (previousNumberOfVariables != static_cast<int>(variablesWithSolution.size()) &&
              m_variablesNames.size() != variablesWithSolution.size());
 }
-
 void SolverUsingSubstitution::addIfSolutionIsCompleteAndCorrect(
     MultipleVariableSolutionSet const& solutionSet, Equations const& equations) {
-    if (m_variablesNames == solutionSet.getVariableNames()) {
-        if (isSolutionCorrect(solutionSet, equations)) {
+    if (m_variablesNames == solutionSet.getVariableNames()) {        if (isSolutionCorrect(solutionSet, equations)) {
             m_solutionsWithAllVariables.emplace_back(solutionSet);
             setAsCompleteSolution();
         }
